taskmanager/requirements.txt
asgiref
Django
django-simple-history
djangorestframework
sqlparse
